+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - add `uninstall` CLI command
 - add `--use-gradient-checkpointing` CLI argument
 - add `--overwrite-repo` CLI argument
+- add `--target-repo` CLI argument for pushes
 
 ## 2.4.0
 - upgrade all dependencies